+++ Sources/StreamChatUI/Deprecations.swift
@@ -5,3 +5,14 @@
 import StreamChat
 
 /// - NOTE: Deprecations of the next major release.
+
+@available(*, deprecated, renamed: "VideoLoading")
+public typealias VideoPreviewLoader = VideoLoading
+
+public extension Components {
+    @available(*, deprecated, renamed: "videoLoader")
+    var videoPreviewLoader: VideoLoading {
+        get { videoLoader }
+        set { videoLoader = newValue }
+    }
+}